Tests/LibWeb/Text/input/css/registered-property-invalid-initial-value.html

<!doctype html>
<style>
    @property --foo {
        syntax: "*";
        inherits: true;
        initial-value: ];
    }
</style>
<script src="../include.js"></script>
<script>
    test(() => {
        println(document.styleSheets[0].cssRules[0].cssText);
    });
</script>
