Tests/LibWeb/Text/expected/css/registered-property-invalid-initial-value.txt

@property --foo { syntax: "*"; inherits: true; }
